switch to dtslint, tweak type, add test
